Extract import/export logic from UsdMenu
