+++ package/com.unity.formats.usd/Editor/Scripts/Behaviors/UsdMenu.cs
@@ -21,50 +21,8 @@ using USD.NET.Unity;
 
 namespace Unity.Formats.USD
 {
-    public class UsdMenu : MonoBehaviour
+    public class UsdMenu
     {
-        public static Scene InitForSave(string defaultName, string fileExtension = "usd,usda,usdc")
-        {
-            var filePath = EditorUtility.SaveFilePanel("Export USD File", "", defaultName, fileExtension);
-
-            if (filePath == null || filePath.Length == 0)
-            {
-                return null;
-            }
-
-            var fileDir = Path.GetDirectoryName(filePath);
-
-            if (!Directory.Exists(fileDir))
-            {
-                var di = Directory.CreateDirectory(fileDir);
-                if (!di.Exists)
-                {
-                    Debug.LogError("Failed to create directory: " + fileDir);
-                    return null;
-                }
-            }
-
-            InitUsd.Initialize();
-            var scene = Scene.Create(filePath);
-            scene.Time = 0;
-            scene.StartTime = 0;
-            scene.EndTime = 0;
-            return scene;
-        }
-
-        public static Scene InitForOpen()
-        {
-            string path = EditorUtility.OpenFilePanel("Import USD File", "", "usd,usda,usdc,abc");
-            if (path == null || path.Length == 0)
-            {
-                return null;
-            }
-
-            InitUsd.Initialize();
-            var stage = pxr.UsdStage.Open(path, pxr.UsdStage.InitialLoadSet.LoadNone);
-            return Scene.Open(stage);
-        }
-
         [MenuItem("USD/Export Selected with Children", true)]
         static bool EnableMenuExportSelectedWithChildren()
         {
@@ -74,7 +32,7 @@ namespace Unity.Formats.USD
         [MenuItem("USD/Export Selected with Children", priority = 50)]
         static void MenuExportSelectedWithChildren()
         {
-            ExportSelected(BasisTransformation.SlowAndSafe);
+            ExportHelpers.ExportSelected(BasisTransformation.SlowAndSafe);
         }
 
         [MenuItem("USD/Export Transform Overrides", true)]
@@ -87,7 +45,7 @@ namespace Unity.Formats.USD
         static public void MenuExportTransforms()
         {
             var root = Selection.activeGameObject.GetComponentInParent<UsdAsset>();
-            var overs = InitForSave(Path.GetFileNameWithoutExtension(root.usdFullPath) + "_overs");
+            var overs = ExportHelpers.InitForSave(Path.GetFileNameWithoutExtension(root.usdFullPath) + "_overs");
             root.ExportOverrides(overs);
         }
 
@@ -167,188 +125,42 @@ namespace Unity.Formats.USD
     }
 #endif
 
-        static private pxr.SdfPath GetDefaultRoot(Scene scene)
-        {
-            // We can't safely assume the default prim is the model root, because Alembic files will
-            // always have a default prim set arbitrarily.
-
-            // If there is only one root prim, reference this prim.
-            var children = scene.Stage.GetPseudoRoot().GetChildren().ToList();
-            if (children.Count == 1)
-            {
-                return children[0].GetPath();
-            }
-
-            // Otherwise there are 0 or many root prims, in this case the best option is to reference
-            // them all, to avoid confusion.
-            return pxr.SdfPath.AbsoluteRootPath();
-        }
-
-        static void ExportSelected(BasisTransformation basisTransform,
-            string fileExtension = "usd,usda,usdc",
-            bool exportMonoBehaviours = false)
-        {
-            Scene scene = null;
-
-            foreach (GameObject go in Selection.gameObjects)
-            {
-                if (scene == null)
-                {
-                    scene = InitForSave(go.name, fileExtension);
-                    if (scene == null)
-                    {
-                        return;
-                    }
-                }
-
-                try
-                {
-                    SceneExporter.Export(go, scene, basisTransform,
-                        exportUnvarying: true, zeroRootTransform: false,
-                        exportMonoBehaviours: exportMonoBehaviours);
-                }
-                catch (System.Exception ex)
-                {
-                    Debug.LogException(ex);
-                    continue;
-                }
-            }
-
-            if (scene != null)
-            {
-                scene.Save();
-                scene.Close();
-            }
-        }
 
         [MenuItem("USD/Import as GameObjects", priority = 0)]
         public static void MenuImportAsGameObjects()
         {
-            var scene = InitForOpen();
+            var scene = ImportHelpers.InitForOpen();
             if (scene == null)
             {
                 return;
             }
 
-            ImportSceneAsGameObject(scene);
+            ImportHelpers.ImportSceneAsGameObject(scene);
             scene.Close();
-        }
-
-        public static GameObject ImportSceneAsGameObject(Scene scene, SceneImportOptions importOptions = null)
-        {
-            string path = scene.FilePath;
-
-            // Time-varying data is not supported and often scenes are written without "Default" time
-            // values, which makes setting an arbitrary time safer (because if only default was authored
-            // the time will be ignored and values will resolve to default time automatically).
-            scene.Time = 1.0;
-
-            if (importOptions == null)
-            {
-                importOptions = new SceneImportOptions();
-                importOptions.changeHandedness = BasisTransformation.SlowAndSafe;
-                importOptions.materialImportMode = MaterialImportMode.ImportDisplayColor;
-                importOptions.usdRootPath = GetDefaultRoot(scene);
-            }
-
-            GameObject root = new GameObject(GetObjectName(importOptions.usdRootPath, path));
-
-            if (Selection.gameObjects.Length > 0)
-            {
-                root.transform.SetParent(Selection.gameObjects[0].transform);
-            }
-
-            try
-            {
-                UsdToGameObject(root, scene, importOptions);
-                return root;
-            }
-            catch (SceneImporter.ImportException)
-            {
-                GameObject.DestroyImmediate(root);
-                return null;
-            }
         }
 
         [MenuItem("USD/Import as Prefab", priority = 1)]
         public static void MenuImportAsPrefab()
         {
-            var scene = InitForOpen();
+            var scene = ImportHelpers.InitForOpen();
             if (scene == null)
             {
                 return;
             }
 
-            string path = scene.FilePath;
-
-            // Time-varying data is not supported and often scenes are written without "Default" time
-            // values, which makes setting an arbitrary time safer (because if only default was authored
-            // the time will be ignored and values will resolve to default time automatically).
-            scene.Time = 1.0;
-
-            var importOptions = new SceneImportOptions();
-            importOptions.projectAssetPath = GetSelectedAssetPath();
-            importOptions.changeHandedness = BasisTransformation.SlowAndSafe;
-            importOptions.materialImportMode = MaterialImportMode.ImportDisplayColor;
-            importOptions.usdRootPath = GetDefaultRoot(scene);
-
-            var invalidChars = Path.GetInvalidFileNameChars();
-            var prefabName = string.Join("_", GetPrefabName(path).Split(invalidChars,
-                System.StringSplitOptions.RemoveEmptyEntries)).TrimEnd('.');
-            string prefabPath = importOptions.projectAssetPath + prefabName + ".prefab";
-            prefabPath = AssetDatabase.GenerateUniqueAssetPath(prefabPath);
-            string clipName = Path.GetFileNameWithoutExtension(path);
-
-            var go = new GameObject(GetObjectName(importOptions.usdRootPath, path));
-            try
-            {
-                UsdToGameObject(go, scene, importOptions);
-                SceneImporter.SavePrefab(go, prefabPath, clipName, importOptions);
-            }
-            finally
-            {
-                GameObject.DestroyImmediate(go);
-                scene.Close();
-            }
+            ImportHelpers.ImportAsPrefab(scene);
         }
 
         [MenuItem("USD/Import as Timeline Clip", priority = 2)]
         public static void MenuImportAsTimelineClip()
         {
-            var scene = InitForOpen();
+            var scene = ImportHelpers.InitForOpen();
             if (scene == null)
             {
                 return;
             }
 
-            string path = scene.FilePath;
-
-            var invalidChars = Path.GetInvalidFileNameChars();
-            var prefabName = string.Join("_", GetPrefabName(path).Split(invalidChars,
-                System.StringSplitOptions.RemoveEmptyEntries)).TrimEnd('.');
-
-            string prefabPath = GetSelectedAssetPath() + prefabName + ".prefab";
-            prefabPath = AssetDatabase.GenerateUniqueAssetPath(prefabPath);
-            string clipName = Path.GetFileNameWithoutExtension(path);
-
-            var importOptions = new SceneImportOptions();
-            importOptions.projectAssetPath = GetSelectedAssetPath();
-            importOptions.changeHandedness = BasisTransformation.FastWithNegativeScale;
-            importOptions.materialImportMode = MaterialImportMode.ImportDisplayColor;
-            importOptions.usdRootPath = GetDefaultRoot(scene);
-
-            var go = new GameObject(GetObjectName(importOptions.usdRootPath, path));
-
-            try
-            {
-                // Ensure we have at least one GameObject with the import settings.
-                XformImporter.BuildSceneRoot(scene, go.transform, importOptions);
-                SceneImporter.SavePrefab(go, prefabPath, clipName, importOptions);
-            }
-            finally
-            {
-                GameObject.DestroyImmediate(go);
-            }
+            ImportHelpers.ImportAsTimelineClip(scene);
         }
 
         [MenuItem("USD/Unload Subtree", true)]
@@ -417,70 +229,6 @@ namespace Unity.Formats.USD
                 Debug.LogWarning("No USD payloads found in subtree.");
                 return;
             }
-        }
-
-        /// <summary>
-        /// Returns the selected object path or "Assets/" if no object is selected.
-        /// </summary>
-        public static string GetSelectedAssetPath()
-        {
-            Object[] selectedAsset = Selection.GetFiltered(typeof(Object), SelectionMode.Assets);
-            foreach (Object obj in selectedAsset)
-            {
-                var path = AssetDatabase.GetAssetPath(obj.GetInstanceID());
-                if (string.IsNullOrEmpty(path))
-                {
-                    continue;
-                }
-
-                if (File.Exists(path))
-                {
-                    path = Path.GetDirectoryName(path);
-                }
-
-                if (!path.EndsWith("/"))
-                {
-                    path += "/";
-                }
-
-                return path;
-            }
-
-            return "Assets/";
-        }
-
-        public static GameObject UsdToGameObject(GameObject parent,
-            Scene scene,
-            SceneImportOptions importOptions)
-        {
-            try
-            {
-                SceneImporter.ImportUsd(parent, scene, new PrimMap(), importOptions);
-            }
-            finally
-            {
-                scene.Close();
-            }
-
-            return parent;
-        }
-
-        private static string GetObjectName(pxr.SdfPath rootPrimName, string path)
-        {
-            return pxr.UsdCs.TfIsValidIdentifier(rootPrimName.GetName())
-                ? rootPrimName.GetName()
-                : GetObjectName(path);
-        }
-
-        private static string GetObjectName(string path)
-        {
-            return UnityTypeConverter.MakeValidIdentifier(Path.GetFileNameWithoutExtension(path));
-        }
-
-        private static string GetPrefabName(string path)
-        {
-            var fileName = GetObjectName(path);
-            return fileName + "_prefab";
         }
     }
 }
