+++ package/com.unity.formats.usd/Runtime/Scripts/IO/Scene/ExportHelpers.cs.meta
@@ -1,0 +1,3 @@
+﻿fileFormatVersion: 2
+guid: ff4b0fbbfd3e41f782b9e9f03f813653
+timeCreated: 1613596650
