+++ package/com.unity.formats.usd/Runtime/Scripts/IO/Scene/ImportHelpers.cs.meta
@@ -1,0 +1,3 @@
+﻿fileFormatVersion: 2
+guid: 5e226d4ad64f4497ab11ca8ff7ec3e54
+timeCreated: 1613595700
